+++ src/miner/utils/include/Hstring.hh
@@ -23,7 +23,7 @@ public:
 
   Hstring(std::string s, Stype t, expression::Proposition **pp = nullptr);
 
-  Hstring(std::string s, Stype t, int **intv = nullptr);
+  Hstring(std::string s, Stype t, size_t **intv = nullptr);
   
   Hstring();
 
@@ -57,7 +57,7 @@ public:
   expression::Proposition **_pp;
   int _offset;
   std::string _sep;
-  int ** _intv;
+  size_t ** _intv;
 
 private:
   std::vector<Hstring> _append;
